Refactor: split WS layer
