+++ src/main/java/glue/Gachi_Sanchaek/walkLocation/repository/LocationRepository.java
@@ -1,6 +1,6 @@
-package glue.Gachi_Sanchaek.walk.repository;
+package glue.Gachi_Sanchaek.walkLocation.repository;
 
-import glue.Gachi_Sanchaek.walk.entity.WalkLocation;
+import glue.Gachi_Sanchaek.walkLocation.entity.WalkLocation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LocationRepository extends JpaRepository<WalkLocation,Long> {
